refactor: fallback 속성 추가, 배럴파일 생성
