+++ app/components/common/image-with-fallback.tsx
@@ -2,28 +2,26 @@
 
 import { useState } from "react";
 import Image, { ImageProps, StaticImageData } from "next/image";
-import FallbackHero from '@images/fallback-hero.png';
-import FallbackFestival from '@images/fallback-festival.png';
 
 interface ImageWithFallbackProps extends Omit<ImageProps, 'src'> {
   src: string;
-  variant: 'hero' | 'festival';
+  fallback: StaticImageData | string;
 }
 
 export default function ImageWithFallback({
   src,
-  variant,
+  fallback,
   onError,
   ...props
 }: ImageWithFallbackProps) {
-  const [imgSrc, setImgSrc] = useState<string | StaticImageData>(src);
+  const [imgSrc, setImgSrc] = useState<StaticImageData | string>(src);
   return (
     <Image
       {...props}
       src={imgSrc}
       alt={props.alt}
       onError={() => {
-        setImgSrc(variant === 'hero' ? FallbackHero : FallbackFestival)
+        setImgSrc(fallback);
       }}
     />
   );
